# lib/api-client-react/src/generated/api.schemas.ts
// @ts-nocheck
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dayar Al-Ahlam Real Estate API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ApartmentStatus = typeof ApartmentStatus[keyof typeof ApartmentStatus];


export const ApartmentStatus = {
  available: 'available',
  unavailable: 'unavailable',
} as const;

export interface Apartment {
  id: string;
  apartmentNumber: string;
  title: string;
  city: string;
  district: string;
  address: string;
  description: string;
  rooms: number;
  bathrooms: number;
  area: number;
  priceDay: number;
  priceWeek: number;
  priceMonth: number;
  status: ApartmentStatus;
  images: string[];
  /** @nullable */
  video?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ApartmentList {
  data: Apartment[];
  total: number;
  page: number;
  limit: number;
}

export type ApartmentInputStatus = typeof ApartmentInputStatus[keyof typeof ApartmentInputStatus];


export const ApartmentInputStatus = {
  available: 'available',
  unavailable: 'unavailable',
} as const;

export interface ApartmentInput {
  apartmentNumber: string;
  title: string;
  city: string;
  district: string;
  address: string;
  description: string;
  rooms: number;
  bathrooms: number;
  area: number;
  priceDay: number;
  priceWeek: number;
  priceMonth: number;
  status?: ApartmentInputStatus;
  images?: string[];
  video?: string;
}

export type ApartmentUpdateStatus = typeof ApartmentUpdateStatus[keyof typeof ApartmentUpdateStatus];


export const ApartmentUpdateStatus = {
  available: 'available',
  unavailable: 'unavailable',
} as const;

export interface ApartmentUpdate {
  apartmentNumber?: string;
  title?: string;
  city?: string;
  district?: string;
  address?: string;
  description?: string;
  rooms?: number;
  bathrooms?: number;
  area?: number;
  priceDay?: number;
  priceWeek?: number;
  priceMonth?: number;
  status?: ApartmentUpdateStatus;
  images?: string[];
  video?: string;
}

export type StatusUpdateStatus = typeof StatusUpdateStatus[keyof typeof StatusUpdateStatus];


export const StatusUpdateStatus = {
  available: 'available',
  unavailable: 'unavailable',
} as const;

export interface StatusUpdate {
  status: StatusUpdateStatus;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AuthToken {
  success: boolean;
  message: string;
}

export interface AdminInfo {
  username: string;
  authenticated: boolean;
}

export interface MessageResponse {
  message: string;
}

export interface PlatformStats {
  total: number;
  available: number;
  unavailable: number;
  cities: number;
  recentApartments?: Apartment[];
}

export interface CityCount {
  city: string;
  count: number;
}

export interface UploadResponse {
  url: string;
}

export interface UploadImageRequest {
  image?: Blob;
}

export type ListApartmentsParams = {
search?: string;
city?: string;
district?: string;
status?: ListApartmentsStatus;
minPrice?: number;
maxPrice?: number;
sortBy?: ListApartmentsSortBy;
page?: number;
limit?: number;
};

export type ListApartmentsStatus = typeof ListApartmentsStatus[keyof typeof ListApartmentsStatus];


export const ListApartmentsStatus = {
  available: 'available',
  unavailable: 'unavailable',
} as const;

export type ListApartmentsSortBy = typeof ListApartmentsSortBy[keyof typeof ListApartmentsSortBy];


export const ListApartmentsSortBy = {
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  date_asc: 'date_asc',
  date_desc: 'date_desc',
} as const;