# lib/api-zod/src/generated/types/adminInfo.ts
// @ts-nocheck
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dayar Al-Ahlam Real Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminInfo {
  username: string;
  authenticated: boolean;
}