# lib/api-zod/src/generated/types/apartmentInput.ts
// @ts-nocheck
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dayar Al-Ahlam Real Estate API
 * OpenAPI spec version: 0.1.0
 */
import type { ApartmentInputStatus } from './apartmentInputStatus';

export interface ApartmentInput {
  apartmentNumber: string;
  title: string;
  city: string;
  district: string;
  address: string;
  description: string;
  rooms: number;
  bathrooms: number;
  area: number;
  priceDay: number;
  priceWeek: number;
  priceMonth: number;
  status?: ApartmentInputStatus;
  images?: string[];
  video?: string;
}