# lib/api-zod/src/generated/types/apartmentList.ts
// @ts-nocheck
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dayar Al-Ahlam Real Estate API
 * OpenAPI spec version: 0.1.0
 */
import type { Apartment } from './apartment';

export interface ApartmentList {
  data: Apartment[];
  total: number;
  page: number;
  limit: number;
}