# lib/api-zod/src/generated/types/apartmentUpdate.ts
// @ts-nocheck
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dayar Al-Ahlam Real Estate API
 * OpenAPI spec version: 0.1.0
 */
import type { ApartmentUpdateStatus } from './apartmentUpdateStatus';

export interface ApartmentUpdate {
  apartmentNumber?: string;
  title?: string;
  city?: string;
  district?: string;
  address?: string;
  description?: string;
  rooms?: number;
  bathrooms?: number;
  area?: number;
  priceDay?: number;
  priceWeek?: number;
  priceMonth?: number;
  status?: ApartmentUpdateStatus;
  images?: string[];
  video?: string;
}